# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LankaJobs mobile, admin, authentication, and moderation API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthConfig {
  googleEnabled: boolean;
  phoneEnabled: boolean;
}

export interface PhoneRequest {
  /** @minLength 7 */
  phone: string;
}

export type PhoneVerification = PhoneRequest & {
  /** @pattern ^[0-9]{6}$ */
  code: string;
};

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
} as const;

export interface User {
  id: string;
  email?: string | null;
  phone?: string | null;
  name: string;
  role: UserRole;
  status: UserStatus;
}

export interface AuthResponse {
  accessToken: string;
  expiresAt: string;
  user: User;
}

export interface Profile {
  userId?: string;
  location?: string;
  interests?: string[];
  experience?: string;
  education?: string;
  skills?: string[];
  employmentType?: string;
  workPreference?: string;
  workMode?: string;
}

export interface MeResponse {
  user?: User;
  profile?: Profile;
}

export type ProfileInput = Profile;

export interface Category {
  id?: string;
  name?: string;
  slug?: string;
  icon?: string;
  isActive?: boolean;
}

export interface CategoryList {
  data?: Category[];
}

export interface JobInput {
  title: string;
  company: string;
  description: string;
  categoryId?: string;
  location: string;
  employmentType: string;
  workMode: string;
  salary?: string;
  experience?: string;
  education?: string;
  skills?: string[];
  applicationEmail?: string;
  applicationUrl?: string;
  closingDate?: string;
}

/**
 * Invalid request
 */
export type InvalidRequestResponse = void;

/**
 * Authentication required
 */
export type UnauthorizedResponse = void;

/**
 * Insufficient permissions
 */
export type ForbiddenResponse = void;

/**
 * Resource not found
 */
export type NotFoundResponse = void;

/**
 * Provider configuration missing
 */
export type ProviderUnavailableResponse = void;

export type GoogleAuthCallbackParams = {
code?: string;
state?: string;
};

export type ListJobsParams = {
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: number;
search?: string;
category?: string;
location?: string;
employmentType?: string;
workMode?: string;
};

export type ModerateJobBodyStatus = typeof ModerateJobBodyStatus[keyof typeof ModerateJobBodyStatus];


export const ModerateJobBodyStatus = {
  PUBLISHED: 'PUBLISHED',
  REJECTED: 'REJECTED',
  SUSPENDED: 'SUSPENDED',
} as const;

export type ModerateJobBody = {
  status: ModerateJobBodyStatus;
  rejectionReason?: string;
};