# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: LankaJobs mobile, admin, authentication, and moderation API
servers:
  - url: /api
...
paths:
...
  /auth/phone/request:
    post:
      operationId: requestPhoneOtp
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneRequest"
      responses:
        "202": { description: Verification code sent }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "503": { $ref: "#/components/responses/ProviderUnavailable" }
  /auth/phone/verify:
    post:
      operationId: verifyPhoneOtp
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneVerification"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/google/start:
    get:
      operationId: startGoogleAuth
      tags: [auth]
      responses:
        "302": { description: Redirect to Google }
        "503": { $ref: "#/components/responses/ProviderUnavailable" }
  /auth/google/callback:
    get:
      operationId: googleAuthCallback
      tags: [auth]
      parameters:
        - { name: code, in: query, required: false, schema: { type: string } }
        - { name: state, in: query, required: false, schema: { type: string } }
      responses:
        "302": { description: Redirect to configured app callback }
        "400": { $ref: "#/components/responses/InvalidRequest" }
  /auth/me:
    get:
      operationId: currentUser
      tags: [auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user and profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /admin/dashboard:
    get:
      operationId: adminDashboard
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Platform statistics }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/users:
    get:
      operationId: adminUsers
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: User list }
  /admin/jobs:
    get:
      operationId: adminJobs
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Job moderation list }
  /admin/jobs/{id}/moderate:
    post:
      operationId: moderateJob
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [PUBLISHED, REJECTED, SUSPENDED] }
                rejectionReason: { type: string }
      responses:
        "200": { description: Moderated job }
  /admin/categories:
    get: { operationId: adminCategories, tags: [admin], security: [{ bearerAuth: [] }], responses: { "200": { description: Category list } } }
    post: { operationId: createCategory, tags: [admin], security: [{ bearerAuth: [] }], responses: { "201": { description: Created category } } }
  /admin/reports:
    get: { operationId: adminReports, tags: [admin], security: [{ bearerAuth: [] }], responses: { "200": { description: Report list } } }
  /admin/settings:
    get: { operationId: adminSettings, tags: [admin], security: [{ bearerAuth: [] }], responses: { "200": { description: Platform settings } } }
    put: { operationId: updateAdminSettings, tags: [admin], security: [{ bearerAuth: [] }], responses: { "200": { description: Updated settings } } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    InvalidRequest: { description: Invalid request }
    Unauthorized: { description: Authentication required }
    Forbidden: { description: Insufficient permissions }
...
    ProviderUnavailable: { description: Provider configuration missing }
  schemas:
...
    PhoneRequest:
      type: object
      required: [phone]
      properties: { phone: { type: string, minLength: 7 } }
    PhoneVerification:
      allOf:
        - $ref: "#/components/schemas/PhoneRequest"
        - type: object
          required: [code]
          properties: { code: { type: string, pattern: "^[0-9]{6}$" } }
    AuthResponse:
      type: object
      required: [accessToken, expiresAt, user]
      properties:
        accessToken: { type: string }
        expiresAt: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
    MeResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        profile: { $ref: "#/components/schemas/Profile" }
    User:
      type: object
      required: [id, name, role, status]
      properties:
        id: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        name: { type: string }
        role: { type: string, enum: [USER, ADMIN, SUPER_ADMIN] }
        status: { type: string, enum: [ACTIVE, SUSPENDED] }
    Profile:
      type: object
      properties:
        userId: { type: string }
        location: { type: string }
        interests: { type: array, items: { type: string } }
        experience: { type: string }
        education: { type: string }
        skills: { type: array, items: { type: string } }
        employmentType: { type: string }
        workPreference: { type: string }
        workMode: { type: string }
...