# lib/api-zod/src/generated/types/phoneVerification.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LankaJobs mobile, admin, authentication, and moderation API
 * OpenAPI spec version: 1.0.0
 */
import type { PhoneRequest } from './phoneRequest';

export type PhoneVerification = PhoneRequest & {
  /** @pattern ^[0-9]{6}$ */
  code: string;
};