# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LankaJobs mobile, admin, authentication, and moderation API
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: string;
  email?: string | null;
  phone?: string | null;
  name: string;
  role: UserRole;
  status: UserStatus;
}